+++ spec/shelly/app_spec.rb
@@ -276,7 +276,8 @@ describe Shelly::App do
       @app.redeem_code = "foo123"
       attributes = {
         :code_name => "fooo",
-        :redeem_code => "foo123"
+        :redeem_code => "foo123",
+        :organization_name => nil
       }
       @client.should_receive(:create_app).with(attributes).and_return("git_url" => "git@git.shellycloud.com:fooo.git",
         "domains" => %w(fooo.shellyapp.com))